Implement MediaCapabilities API encodingInfo for WebRTC
"webrtc" was recently added as an MediaEncodingType to the
MediaCapabilities API, see
https://w3c.github.io/media-capabilities/

"webrtc" replaces the deprecated encoding type "transmission".

This CL implements support for responding to supported and powerEfficient.
smooth is currently hardcoded to be true given that the configuration
is supported.

The feature is implemented behind the blink runtime enabled flag
MediaCapabilitiesWebRtc

Bug: chromium:1187565
Change-Id: I98d8d971304194d93aa436b8821ea665dfdab9f7